+++ auth-db/queries/users/index.ts
@@ -2,3 +2,4 @@ export * from "./create.ts";
 export * from "./get-all.ts";
 export * from "./get-by-email.ts";
 export * from "./get-by-id.ts";
+export * from "./update-last-login.ts";
